refactor(backend): AiServiceをSensitiveMediaDetectionServiceにリネーム
NSFW推論を本体で行わなくなり、外部サービス (sensitive-detector) への
アダプタになったため、AiServiceという名前が実態と合わなくなっていた。
ログドメインも `ai` から `sensitive-media-detection` に変更する。
