+++ packages/backend/test/unit/FileInfoService.ts
@@ -14,7 +14,7 @@ import { mockDeep } from 'vitest-mock-extended';
 import { GlobalModule } from '@/GlobalModule.js';
 import { FileInfo, FileInfoService } from '@/core/FileInfoService.js';
 //import { DI } from '@/di-symbols.js';
-import { AiService } from '@/core/AiService.js';
+import { SensitiveMediaDetectionService } from '@/core/SensitiveMediaDetectionService.js';
 import { LoggerService } from '@/core/LoggerService.js';
 import type { TestingModule } from '@nestjs/testing';
 
@@ -41,13 +41,13 @@ describe('FileInfoService', () => {
 				GlobalModule,
 			],
 			providers: [
-				AiService,
+				SensitiveMediaDetectionService,
 				LoggerService,
 				FileInfoService,
 			],
 		})
 			.useMocker((token) => {
-				//if (token === AiService) {
+				//if (token === SensitiveMediaDetectionService) {
 				//	return {  };
 				//}
 				if (typeof token === 'function') {
